requirements.txt
ruff
Pillow 
sweepai